Move histogram dash/legend helpers to plots-general.tsx, extract more shared plot logic
Moves computeColorItems/computeDashPatterns/manualLegend (generalized
in #2062) out of plots-histogram.tsx into plots-general.tsx, the
shared plot-infrastructure module, so they're reusable by other plot
types. Alongside the move, pulls out three more pieces Histogram had
that were begging to be shared:

- PlotDownloadButton: the screenshot-download icon + createScreenshot
  wiring from HistogramSettings.
- axisAndGrid: the transpose-based axis/grid mark-constructor picker,
  used twice in this file (xAxis and yAxis each had their own copy).
- multiSeriesTipTitle + transposeAwareTip: the "prefix + per-series
  name:value lines, or a bare single value" tooltip title logic and
  its Plot.tip/pointerX/pointerY wiring.

No behavior change: same marks, same tooltip text, same download
filename. Verified with tsc, eslint, and knip (no new unused-export
findings).
